Refactorying to support authentication on different pages.
Given that our cloud application follows a model in which users such as neurologists &
other health care specialists will be explicitly invited to join into this cloud application.

Thus, the login (sign-in), sign-up & confirm sign-up phases are split to distinct pages so
that we can adapt it accordingly. Eventually, the cloud application will have users having
user adminstrative privileges such as the registering (sign-up) a new user and/or possibly
allowing potential users (those that have requested a sign-up) to join.

Much of this process is generic and which the backend by Amazon's AWS services. But, this
design and implementation shall by simple to adapt to use another cloud infrastructure or
one that could be customized for a given insitution should they have a user management
solution that they rely upon.

Additionally, removed the Amazon Amplify tutorial code, documentation, licensing information
 and others artifacts which aren't appropriate for this cloud application.