Horizon-7: Selecting a recording session to be viewed.
Added the U/I scaffolding & styling for a table (using AG Grid) that displays all of the EEG recording sessions allowing for the user to select which EEG recording session to view.

The next step will be to adapt from a mock example using cars to be replaced by EEG recording sessions from the application's recordings table which is stored in the cloud (i.e. AWS DynamoDB).