Horizon-7: Selecting a recording session to be viewed.
Adapted the presentation of the grid to reflect recording sessions.

Recording sessions which are ongoing a live-feed are placed at the top and the cell is represented by an icon to highlight that it's live-feed.

The start/finish times displays the date & time portions, but filtering hasn't been adapted to allow for specifying time at this point.

Next steps are:
- Feeding the grid with the actual records from the recordings table.
- Increasing the displayable area so that the grid can be larger.
- Find a Live icon that better suited to display in grid's dimensions with respects to height of about 25 pixels

For another story
- Allow users to see the recording sessions' date & time relative to the time-zone of where done.