Horizon-9: Check if user is already logged in from a previous session.
AWS Amplify stores the authenticated user credentials in the browser's cache for a given period of time or until they explicitly logged-out.

In the case, in which the user has logged in and uses the application, then to exit the browser, then when opening the cloud application the user can continue using the application without having to login.

This is to fix the situation in which the login of a user account fails since a user is already logged in.