Horizon-10: Adding the ability to load recording sessions into front end.
The code is in place so that the AWS Amplify Gen 2 code fetches (lists) the recording sessions from the recordings table, but for some reason it's throwing an exception relating to authentication from AWS layer.

Adapted for the refresh button invokes the same fetch recordings as invoked when loading the page scenario.

Modified the Schema model such that the recordings model requires authentication users. Eventually, we'll improve so that records can be read iff the user is in the correct group(s) for that record.