Horizon-10: Adding the ability to load recording sessions into front end.
Added a function and interface explicit of the loading of the recording sessions such that the sessions selection page isn't burdened with the details of this implementation.

Eventually convert to an interface pattern with dependency injection making mocking and extending to different cloud provider & databases will be possible.

Introduced several functions, such as converting a text to a Date type, to reduce that complexity from the sessions selection page's implementation.