HORIZON-11: Incorporate EEG chart view
This phase I've ported the code necessary to fetch the EEG readings from the AWS S3 bucket.

This includes the adapting code to extract authentication session credentials from AWS Amplify and passing it to the AWS SDK layer when establishing a connection to the AWS S3 bucket.

Also includes the code for the ChartLoader to fetches the EEG readings from the AWS S3 bucket and to post a message so that the chart object can load it.