HORIZON-11: Incorporate EEG chart view
Added a skeleton session view page so that upon selecting a recording session from the grid it present this page.

This session view page proceeds using the the user's credentials to enumerate a fixed S3 bucket and folder and proceeds to fetch the contents of the first file.

At present, the approach used is to apply the policy to the S3 bucket. Eventually, the better approach will be to assigned IAM roles as to coordinate the users from the identity pool to the appropriate S3 buckets which they are allowed to fetch the recording session details.