HORIZON-11 Continued work in progress.
This builds and getting close to the point in which the controllers (timeline, chart) and their views can be constructed.

The chart controller which consists of a:
  o data source (which is a web-worker)
  o view (which is Sci-Chart)

The timeline controller which consists of a:
  o video controller(s)
  o chart controller(s)

I've put plural since conceptually, there could be one to many  controllers that need to be synchronized in which 1 controller has taken control of the timeline which depends on which view that the user is interacting with.