HORIZON-11 Continued work in progress.
Added overview, the annotation image appears at the top and the annotations top chart does have labels.

The timeline controller with the chart view appears to be working.

The overview isn't working as expected since it's span is much shorter than expected thus cannot scroll to the end of the recording.