HORIZON-12 Restoring support for video playback.
Restoring support for video playback. At this moment using react-hls-player. However, this won't suffice since it demands that the src is known prior to loading the page.

Our situation is that the record is loaded asynchronously.

We might conceive of a means in which the recording details are passed to the page as properties. However, I prefer the approach of passing the recording id to the page.

Thus, making the page autonomous where staff can copy & paste the link of a given recording session (provided that they are logged-in).