HORIZON-12 Restoring support for video playback.
No longer using react-hls-player since the source isn't known until loading the recording session details which is done asynchronously.

Additionally, put the foundation for a video- controller that coordinates with a timeline-controller. We'll be able to leverage later on.

What remains is to load the chart & video with the values read from the record of a recording session.