HORIZON-12 Restoring support for video playback.
Adapted the model for the recording session so that a folder support different storage solutions from AWS, Azure, etc... Additionally, in the case of AWS include the region.

The next step will be putting those details into the recording session's record by the VEEGix8 app and then adapting the lines of code to read the kind, region and URL to the storage folder.

Discovered how to get at the type definitions from AWS Amplify. Thus, added the code to transform from the data model in GraphQL (AWS DynamoDB) into an application representation that's more practical to work with.

Code cleanup.
  - Considerably due to introduction of the storage.Path interface to reduce passing bucket, folder & region.